% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Node-class}
\alias{Node}
\alias{Node-class}
\title{A Reference Class to represent a node of a game tree.}
\description{
A Reference Class to represent a node of a game tree.
}
\section{Fields}{

\describe{
\item{\code{properties}}{Character vector.}

\item{\code{children}}{Character vector. Children nodes of the current node.}
}}

